[Justice Counts] Add last_visit and visit_this_week to CSG Data Pull (Recidiviz/recidiviz-data#26091)
## Description of the change

On 11/6/23, Lili noticed that the Scoreboard tab in the most recent data
pull (from 11/3/23) had a suspiciously large value for the `Number Of
Agencies Logged In Last Week` metric.

![image](https://github.com/Recidiviz/recidiviz-data/assets/82831800/2f33f213-feb1-426c-8ac9-c97ba0f6f143)

For `last_login`, we track an agency's most recent login date of any of
its users. The reason that this value is so large is because we
previously did not track logins by user for specific agencies. So if a
user belonged to multiple agencies (for example, a super agency with
many child agencies), if that user logged in to 1 of those agencies,
that login counted towards all of the agencies that the user belongs to.

However, as of Recidiviz/recidiviz-data#25801 and
[Recidiviz/recidiviz-data#1083](https://github.com/Recidiviz/justice-counts/pull/1083), we now
track user visits to all agencies (user/agency pair visits) in our
backend and db. This is true for whenever a user logs in to a given
agency, a user switches agencies via the Dropdown menu, or a user
switches agencies from any page by changing the agency_id in the url.
This also is true for both regular, super, and child agencies.

This PR adds the updated `last_visit` and subsequent `visit_this_week`
and `Number Of Agencies Visited Last Week` to the weekly Justice Counts
Data Pull. These columns will be more accurate than the older
`last_login`, `login_this_week`, and `Number Of Agencies Logged In Last
Week` columns. We will keep the older columns for now as the newer
columns will need some time to populate.

How I Tested: I ran the csg_data_pull locally and it produced the
following sheets
<img width="1840" alt="Screenshot 2023-12-26 at 12 41 56 PM"
src="https://github.com/Recidiviz/recidiviz-data/assets/82831800/88c97dda-bcad-4af1-b8b5-e9245f937bb8">
<img width="1840" alt="Screenshot 2023-12-26 at 12 42 21 PM"
src="https://github.com/Recidiviz/recidiviz-data/assets/82831800/35a68487-a050-4644-a94c-a51428c5f88d">

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#25231

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Michelle Orden <michelleorden@Michelles-MacBook-Pro.local>
Co-authored-by: Michelle Orden <michelleorden@Michelles-MBP.attlocal.net>
GitOrigin-RevId: 89d7a7fd3bd1dc6d7d9f778831d5c4d2587335d7