allow duplicate rows in SupervisionClientEvent table (Recidiviz/recidiviz-data#26509)
## Description of the change

As per
https://recidiviz.slack.com/archives/C03C24ZSK4K/p1703709022570809,
there can be multiple events of the same type for a given client/officer
on the same day, and it seemed based on querying that these can have the
same attributes as well. Based on that we can't set any combination of
our columns to be the primary key, because we might have some rows that
are fully identical. We may decide in the future to assume that if the
attributes are the same that it's actually the same event, but for now,
allow these fully duplicate rows to unblock importing the data into
Cloud SQL.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://github.com/Recidiviz/recidiviz-data/issues/25493

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3d3f2a4729ce2fdafcd3dedc94877af5fab775ce