[US_MI] Incorporate historical modifiers data and fix COMS data incorporation bug (Recidiviz/recidiviz-data#26583)
## Description of the change

This PR revises supervision period ingest such that:
- We now use the `End_Date` column from COMS_Modifiers as
modifier_end_date instead of inferring it from the ALL version of the
table
- Previously, we didn't have the `End_Date` column and also only had
active modifiers so we had to infer end dates
- Add back in the code that ends all supervision level and supervision
officer assignment spans from OMNI on 8/14/2023 (the date of the
migration from OMNI to COMS)
- Previously, we had this in the code, but we removed it in
[this](https://github.com/Recidiviz/recidiviz-data/pull/26090) PR when
we revised it such that ingest always defaulted to using COMS even if it
was before the migration. However, this led to issues in the case where
it was after the migration date, someone's supervision level or officer
assignment ended in COMS, no new supervision level or officer assignment
was added into COMS, and the supervision level or officer assignment was
never closed out in OMNI. This would lead ingest to go back to using the
supervision level or officer assignment from OMNI which is inaccurate.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2a6f537ec91035170cfd778283efef02ccb484de