[Ingest] Unzip raw data files when they have normalized paths (Recidiviz/recidiviz-data#26599)
## Description of the change

We were mistakenly only unzipping raw data zip files when they did not
have normalized file names. This updates to unzip files after their
paths have been normalize so we handle ZIP files whether or not they are
uploaded pre-normalized (i.e. during SFTP). Also updates so the inner
files are normalized with the same date as the outer file.

See discuss in this Slack thread:
https://recidiviz.slack.com/archives/C028X32LRH7/p1704320591585809.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#26221

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8dbf2011831f76bca5de3a7bf63e6006e49e52b4