[US_OZ] add raw data file configs for hunger games file tags (Recidiviz/recidiviz-data#26567)
## Description of the change

Adds raw data configs (yaml & markdown) for the `hunger_games_people`,
`hunger_games_periods`, and `hunger_games_backgrounds` file tags for
`US_OZ`. The raw file are in the
[recidiviz-staging-direct-ingest-state-us-oz](https://console.cloud.google.com/storage/browser/recidiviz-staging-direct-ingest-state-us-oz)
bucket.

Uses Recidiviz/recidiviz-data#15025 as a PR template.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| **Type: Feature** | **non-breaking change that adds functionality** |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

n/a

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 54be7cc9a336d1062eea96b8774731f56bbde505