[Justice Counts] Copy all recidiviz-data code into the JC Docker image. (Recidiviz/recidiviz-data#26590)
## Description of the change

Copy all recidiviz-data code into the JC Docker image.

Context: Recidiviz/recidiviz-data#26586

JC code only relies on a subset of the recidiviz-data directories, so
currently we only copy those into the JC docker image. Recently,
however, some of the utilities we rely on were moved to directories we
don't copy over which caused a dependency error.

This can be solved by copying over all of the recidiviz-data repo to our
docker image instead of just a subset of directories.

It seems like doing this does not notably increase the startup time of
our docker image, and would help avoid these dependency issues in the
future.

## Testing
Started up the docker image in playtesting and verified that it worked
as expected.

Also test built the Docker image locally and then ran `pipenv run
docker-jc` to verify that everything is working.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#26586

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3cf3836ae00834421e41845f5538ba353558b7ad