Fixes bug in custom metric lookml autogeneration using _in_query to compose a SQL query (Recidiviz/recidiviz-data#26719)
## Description of the change

We were previously seeing buggy behavior with the _in_query liquid
condition such that if someone selected a measure in the custom metrics
explore, the liquid SQL query would not add that metric to dynamically
composed query. This was because of a couple syntax errors: 1) should
use `_in_query`, not `in_query`, and 2) statements were referencing
measures without the `_measure` suffix.

Corresponding LookML output can be found
[here](https://github.com/Recidiviz/looker/pull/528).

This change is one prerequisite for proper functionality of existing
custom metrics infra, and will also unblock workflows impact measurement
infrastructure via custom metrics in looker.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#26525

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4121f6d6d0b57c4eb0b37f5df9ec26614df6904c