[Justice Counts] [1/n] bug fix to remove unnecessary replacement of supervision subsystem (Recidiviz/recidiviz-data#26735)
## Description of the change

I ran into a bug while testing the reminder email job in staging where I
was creating a metric definition key incorrectly for supervision
subsystems. This caused the key error that was surfaced in
[sentry](https://recidiviz-inc.sentry.io/issues/4839728718/?alert_rule_id=13502472&alert_timestamp=1704834154809&alert_type=email&environment=production&notification_uuid=e3759aff-1366-4942-bb33-524d12ff3628&project=4504532096516096&referrer=alert_email).
Turns out, I don't have to manually add the metric definitions for the
supervision subsystems when a metric is reported by supervision
subsystems. When a user decides a metric will be reported by supervision
subsystem, we add an agency datapoint for every supervision subsystem
with the `DISAGGREGATED_BY_SUPERVISION_SUBSYSTEMS` context key. Along
with that we add an agency datapoint showing that the metric is enabled
for the supervision subsystem. I had incorrectly assumed that only the
combined supervision metric was marked
`DISAGGREGATED_BY_SUPERVISION_SUBSYSTEMS` and enabled and added a lot of
code that I didn't end up needing. I also updated the unit test to
reflect a the realistic database conditions for the agency datapoints.

Along with that, I tested locally to confirm that this change worked
with the agency that was surfacing the key error by running the script
in staging with these changes and confirming the emails were saved
successfully.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8b1ddb197ea0a238a635e58aca7c6b4e497fe922