[Justice Counts] [2/n] Save email html in GCP bucket when sending reminder emails (Recidiviz/recidiviz-data#26750)
## Description of the change

New feature where we save the html content of the reminder emails that
we will be sending to a GCP bucket. It is important to save this
information so that we can review what we send to agencies if they have
questions and check the content of the emails after launch.

https://github.com/Recidiviz/recidiviz-data/assets/19961693/63eb8aff-91ec-418a-9c9f-3fa30db611db

## Related issues

Closes Recidiviz/recidiviz-data#26749

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 18a8c2282bd55e725aa81461235c8efe241f6567