[Normalization] Don't skip IP normalization when there are no IPs (Recidiviz/recidiviz-data#26767)
## Description of the change

We were skipping all IP normalization logic when there were no input
IPs, which is used to be fine to do because normalization only modified
or deleted existing IPs, but is not a valid approach now that some
custom normalization logic can choose to add IPs based on the existence
of other entities (e.g. add an IP when we see an IN_CUSTODY SP).

This change updates logic to always run the full IP normalization flow
for all people.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to work being done in Recidiviz/recidiviz-data#26502.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: cf690eaee49a53a2103bc6c82a8b3b0e99b153eb